Fixed occasional double-logging issue when logging at the same time from two different threads would cause the two logs to get merged with each other. Separated Log literals and the Logger class into different files. Moved log-related files to a new folder.
